# ems/api/oapi/api.yaml
openapi: 3.0.3
info:
  title: EMS frontend API
  description: An API for accessing configuration page of EEPROM Monitoring Server
  license:
    name: GNU General Public License version 3 or later
  version: 0.1.0
paths:
  /signin:
    get:
      summary: Load signin page
      responses:
        200:
          description: Returns the signin landing page
          $ref: '#/components/responses/Page'
        500:
          description: Internal server error
          $ref: '#/components/responses/PageError'
      security: []
    post:
      summary: Handle login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                login:
                  type: string
                  format: email
                  pattern: '^[a-zA-Z0-9]([-_.a-zA-Z0-9]){1,31}$'
                  minLength: 2
                  maxLength: 32
                password:
                  type: string
                  format: password
              required:
              - login
              - password
      responses:
        200:
          description: Wrong credentials
          $ref: '#/components/responses/Page'
        303:
          description: Login success
          $ref: '#/components/responses/PageRedirect'
        500:
          description: Internal server error
          $ref: '#/components/responses/PageError'
      security: []
...
components:
...
  responses:
    Page:
      description: Load requested page
      content:
        text/html:
          schema:
            type: string
    PageRedirect:
      description: Redirect to page
      headers:
        Location:
          description: Direction of redirect
          schema:
            type: string
            example: /signin
            enum:
            - /
            - /signin
    PageError:
      description: Load requested page with error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: 'Error message'
              errorDetails:
                type: string
                example: 'Error detailed message'
            required:
            - error
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session_token
      in: cookie